Added new module to support LXC containers
The new module will allow users to control LXC containers from ansible.
The module was built for use in LXC >= 1.0 or greater and implements most
of what can be done using the various lxc clients with regards to running
containers. This first module is geared only at managing lxc containers.

The module provides:
  build containers
  destroy containers
  archive containers
  info from a single container
  start / stop / restart containers
  run commands within containers
  add/modify lxc config for a container
  supports backends including LVM