Be consistent in the documentation
The documentation for the `state` field is not very clear.

It says possible values are "installed, uninstalled" and default value is "present"
The examples below alow uses `present` and `absent`.

This patch uses "absent" and "present" instead of "installed" and "uninstalled"

Moreover, this is consistent with other packaging modules, like homebrew itself