Change 'values' to 'record_data' for gcdns module. (#3003)
Using values caused problems while creating an integration playbook as it is a reserved word.  Seeing as this module is not yet released, it's prudent to make this change now.  'record_data' is more descriptive and uses the _data convention that we've established for instances.

No functionality in the module has changed.